#43 使いやすく見やすくするために登録と更新のロジックをカスタムフックス化した。
