+++ src/components/hooks/useAlert.tsx
@@ -1,0 +1,29 @@
+import Router from 'next/router';
+import React from 'react';
+import { TatoeBtnHooksProps } from '../types/types';
+
+export const useAlert = (
+  props: Pick<
+    TatoeBtnHooksProps,
+    'userId' | 'user' | 'title' | 'shortParaphrase' | 'description'
+  >
+) => {
+  const { userId, user, title, shortParaphrase, description } = props;
+
+  const alertRegisterTatoe = (): void => {
+    if (!userId || !user) {
+      return null;
+    }
+    if (title === '' || shortParaphrase === '' || description === '') {
+      alert('入力されていない箇所があります。');
+      return;
+    }
+
+    if (user.userName === '') {
+      alert('ユーザー名を登録して投稿してください。');
+      return;
+    }
+  };
+
+  return { alertRegisterTatoe };
+};
