+++ src/components/hooks/useTatoe.tsx
@@ -1,0 +1,145 @@
+import { Tatoe, TatoeBtnHooksProps } from '../types/types';
+import { useAlert } from './useAlert';
+
+export const useTatoe = (props: TatoeBtnHooksProps) => {
+  const {
+    router,
+    setTatoe,
+    tatoe,
+    userId,
+    user,
+    title,
+    shortParaphrase,
+    description,
+    query_tId,
+    persistAccessToken,
+    createdAt,
+    updatedAt,
+  } = props;
+
+  const handleOnClickCreateTatoe = async (): Promise<void> => {
+    // TODO アラートは出るのにリターンにならない。そのまま遷移して空の１個のリストができてしまう。更新もおなじく。
+    const { alertRegisterTatoe } = useAlert({
+      userId,
+      user,
+      title,
+      shortParaphrase,
+      description,
+    });
+    alertRegisterTatoe();
+
+    // if (!userId || !user) {
+    //   return null;
+    // }
+    // if (title === '' || shortParaphrase === '' || description === '') {
+    //   alert('入力されていない箇所があります。');
+    //   return;
+    // }
+
+    // if (user.userName === '') {
+    //   alert('ユーザー名を登録して投稿してください。');
+    //   return;
+    // }
+    // 例え登録のロジックを書く
+    if (!query_tId) {
+      const res = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/tatoe`, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: `Bearer ${persistAccessToken}`,
+        },
+        body: JSON.stringify({
+          title,
+          shortParaphrase,
+          description,
+        }),
+      });
+      const { data } = await res.json();
+      // フロント側で使えるようにデータ加工
+      const formattedData: Tatoe = {
+        tId: data.id,
+        userId: data.userId,
+        createdAt: data.createdAt,
+        updatedAt: data.updatedAt,
+        title: data.title,
+        description: data.description,
+        shortParaphrase: data.shortParaphrase,
+      };
+      const newTatoe = [formattedData, ...tatoe];
+
+      setTatoe(newTatoe);
+
+      router.push({
+        pathname: '/DashBoard/UserTatoeList',
+      });
+    }
+  };
+
+  const handleOnclickUpdateTatoe = async (): Promise<void> => {
+    const { alertRegisterTatoe } = useAlert({
+      userId,
+      user,
+      title,
+      shortParaphrase,
+      description,
+    });
+    alertRegisterTatoe();
+    // if (!userId || !user) {
+    //   return null;
+    // }
+    // if (title === '' || shortParaphrase === '' || description === '') {
+    //   alert('入力されていない箇所があります。');
+    //   return;
+    // }
+
+    // if (user.userName === '') {
+    //   alert('ユーザー名を登録して投稿してください。');
+    //   return;
+    // }
+    // 例え更新のロジックを書く
+    if (query_tId) {
+      tatoe.map(async (item: Tatoe) => {
+        if (item.tId === query_tId) {
+          const tId = item.tId;
+          const res = await fetch(
+            `${process.env.NEXT_PUBLIC_BACKEND_URL}/tatoe/${tId}`,
+            {
+              method: 'PUT',
+              headers: {
+                'Content-Type': 'application/json',
+                Authorization: `Bearer ${persistAccessToken}`,
+              },
+              body: JSON.stringify({
+                title,
+                shortParaphrase,
+                description,
+                tId,
+                createdAt,
+                updatedAt,
+              }),
+            }
+          );
+          const { data } = await res.json();
+
+          const updatedTatoe = tatoe.map((item: Tatoe) => {
+            if (item.tId === query_tId) {
+              return {
+                tId: item.tId,
+                userId: data.userId,
+                createdAt: data.createdAt,
+                updatedAt: data.updatedAt,
+                title: data.title,
+                description: data.description,
+                shortParaphrase: data.shortParaphrase,
+              };
+            }
+            return item;
+          });
+
+          setTatoe(updatedTatoe);
+        }
+      });
+    }
+  };
+  return { handleOnClickCreateTatoe, handleOnclickUpdateTatoe };
+};
