+++ src/pages/Register/RegisterTatoeParent.tsx
@@ -9,10 +9,10 @@ import { useRouter } from 'next/router';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import { TatoeAtom } from '../../components/utils/atoms/TatoeAtom';
 import { LoginUserAtom } from '../../components/utils/atoms/LoginUserAtom';
-import { Tatoe } from '../../components/types/types';
 import { UpdateTatoeBtn } from './RegisterTatoeChild/UpdateTatoeBtn';
 import { useAuth } from '../../components/hooks/useAuth';
 import { useUserInfo } from '../../components/hooks/useUserInfo';
+import { useTatoe } from '../../components/hooks/useTatoe';
 
 export const RegisterTatoeParent = () => {
   const router = useRouter();
@@ -29,112 +29,20 @@ export const RegisterTatoeParent = () => {
 
   const query_tId = query.tId;
 
-  const handleOnClickCreateTatoe = async (): Promise<string> => {
-    if (!userId || !user) {
-      return null;
-    }
-    if (title === '' || shortParaphrase === '' || description === '') {
-      alert('入力されていない箇所があります。');
-      return;
-    }
-
-    if (user.userName === '') {
-      alert('ユーザー名を登録して投稿してください。');
-      return;
-    }
-    // 例え登録のロジックを書く
-    if (!query_tId) {
-      const res = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/tatoe`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${persistAccessToken}`,
-        },
-        body: JSON.stringify({
-          title,
-          shortParaphrase,
-          description,
-        }),
-      });
-      const { data } = await res.json();
-      // フロント側で使えるようにデータ加工
-      const formattedData: Tatoe = {
-        tId: data.id,
-        userId: data.userId,
-        createdAt: data.createdAt,
-        updatedAt: data.updatedAt,
-        title: data.title,
-        description: data.description,
-        shortParaphrase: data.shortParaphrase,
-      };
-      const newTatoe = [formattedData, ...tatoe];
-
-      setTatoe(newTatoe);
-
-      router.push({
-        pathname: '/DashBoard/UserTatoeList',
-      });
-    }
-  };
-
-  const handleOnclickUpdateTatoe = async (): Promise<string> => {
-    if (!userId || !user) {
-      return null;
-    }
-    if (title === '' || shortParaphrase === '' || description === '') {
-      alert('入力されていない箇所があります。');
-      return;
-    }
-
-    if (user.userName === '') {
-      alert('ユーザー名を登録して投稿してください。');
-      return;
-    }
-    // 例え更新のロジックを書く
-    if (query_tId) {
-      tatoe.map(async (item: Tatoe) => {
-        if (item.tId === query_tId) {
-          const tId = item.tId;
-          const res = await fetch(
-            `${process.env.NEXT_PUBLIC_BACKEND_URL}/tatoe/${tId}`,
-            {
-              method: 'PUT',
-              headers: {
-                'Content-Type': 'application/json',
-                Authorization: `Bearer ${persistAccessToken}`,
-              },
-              body: JSON.stringify({
-                title,
-                shortParaphrase,
-                description,
-                tId,
-                createdAt,
-                updatedAt,
-              }),
-            }
-          );
-          const { data } = await res.json();
-
-          const updatedTatoe = tatoe.map((item: Tatoe) => {
-            if (item.tId === query_tId) {
-              return {
-                tId: item.tId,
-                userId: data.userId,
-                createdAt: data.createdAt,
-                updatedAt: data.updatedAt,
-                title: data.title,
-                description: data.description,
-                shortParaphrase: data.shortParaphrase,
-              };
-            }
-            return item;
-          });
-
-          setTatoe(updatedTatoe);
-        }
-      });
-    }
-  };
+  const { handleOnClickCreateTatoe, handleOnclickUpdateTatoe } = useTatoe({
+    router,
+    tatoe,
+    user,
+    setTatoe,
+    persistAccessToken,
+    userId,
+    title,
+    shortParaphrase,
+    description,
+    query_tId,
+    createdAt,
+    updatedAt,
+  });
 
   return (
     <div className='flex flex-col gap-6'>
